requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
gigachat>=0.1.0
flask>=3.0.0
flask-cors>=4.0.0
